stale-pr-scanner/requirements.txt
PyGithub>=2.1,<3
PyYAML>=6.0,<7
python-dotenv>=1.0,<2
requests>=2.31,<3
